requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
asyncio==3.4.3
aiohttp==3.9.1
supervisor==4.2.5
python-json-logger==2.0.7
tenacity==8.2.3
typing-extensions==4.8.0
colorama==0.4.6
tqdm==4.66.1
